refactor: 비속어 관련 클래스를 common 의 infra 모듈로 이동
